CLN : literalinclude wizard and configure_traits_view_buttons
	modified:   docs/source/traitsui_user_manual/custom_view.rst
	new file:   docs/source/traitsui_user_manual/examples/wizard.py
	modified:   examples/tutorials/doc_examples/examples/configure_traits_view_buttons.py
	deleted:    examples/tutorials/doc_examples/examples/wizard.py
